Fix pasting of X/Twitter links
There was a bug that pasting link from domains ending with x.com was
parsed as a ex-Twitter link. But this was not necessarily a link from
this social network. This commit changes the regex to better detect
proper X links and tell them apart from non-X links.

This is a port of an original PR https://github.com/logseq/logseq/pull/11942